+++ htmlcov/index.html
@@ -56,7 +56,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.6.9">coverage.py v7.6.9</a>,
-            created at 2024-12-20 16:02 +0300
+            created at 2024-12-25 14:58 +0300
         </p>
     </div>
 </header>
@@ -75,7 +75,7 @@
         </thead>
         <tbody>
             <tr class="region">
-                <td class="name left"><a href="z_c68eb0c7512457e4___init___py.html">praktikum\__init__.py</a></td>
+                <td class="name left"><a href="z_a44f0ac069e85531___init___py.html">tests\__init__.py</a></td>
                 <td>0</td>
                 <td>0</td>
                 <td>0</td>
@@ -84,60 +84,51 @@
                 <td class="right" data-ratio="0 0">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_c68eb0c7512457e4_bun_py.html">praktikum\bun.py</a></td>
-                <td>8</td>
+                <td class="name left"><a href="z_a44f0ac069e85531_test_bun_py.html">tests\test_bun.py</a></td>
+                <td>19</td>
                 <td>0</td>
                 <td>0</td>
                 <td>0</td>
                 <td>0</td>
-                <td class="right" data-ratio="8 8">100%</td>
+                <td class="right" data-ratio="19 19">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_c68eb0c7512457e4_burger_py.html">praktikum\burger.py</a></td>
-                <td>27</td>
+                <td class="name left"><a href="z_a44f0ac069e85531_test_burger_py.html">tests\test_burger.py</a></td>
+                <td>33</td>
                 <td>0</td>
                 <td>0</td>
-                <td>4</td>
                 <td>0</td>
-                <td class="right" data-ratio="31 31">100%</td>
+                <td>0</td>
+                <td class="right" data-ratio="33 33">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_c68eb0c7512457e4_database_py.html">praktikum\database.py</a></td>
-                <td>21</td>
+                <td class="name left"><a href="z_a44f0ac069e85531_test_database_py.html">tests\test_database.py</a></td>
+                <td>26</td>
                 <td>0</td>
                 <td>0</td>
                 <td>0</td>
                 <td>0</td>
-                <td class="right" data-ratio="21 21">100%</td>
+                <td class="right" data-ratio="26 26">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_c68eb0c7512457e4_ingredient_py.html">praktikum\ingredient.py</a></td>
-                <td>11</td>
+                <td class="name left"><a href="z_a44f0ac069e85531_test_ingredient_py.html">tests\test_ingredient.py</a></td>
+                <td>25</td>
                 <td>0</td>
                 <td>0</td>
                 <td>0</td>
                 <td>0</td>
-                <td class="right" data-ratio="11 11">100%</td>
-            </tr>
-            <tr class="region">
-                <td class="name left"><a href="z_c68eb0c7512457e4_ingredient_types_py.html">praktikum\ingredient_types.py</a></td>
-                <td>2</td>
-                <td>0</td>
-                <td>0</td>
-                <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="2 2">100%</td>
+                <td class="right" data-ratio="25 25">100%</td>
             </tr>
         </tbody>
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>69</td>
+                <td>103</td>
                 <td>0</td>
                 <td>0</td>
-                <td>4</td>
                 <td>0</td>
-                <td class="right" data-ratio="73 73">100%</td>
+                <td>0</td>
+                <td class="right" data-ratio="103 103">100%</td>
             </tr>
         </tfoot>
     </table>
@@ -149,12 +140,12 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.6.9">coverage.py v7.6.9</a>,
-            created at 2024-12-20 16:02 +0300
+            created at 2024-12-25 14:58 +0300
         </p>
     </div>
     <aside class="hidden">
-        <a id="prevFileLink" class="nav" href="z_c68eb0c7512457e4_ingredient_types_py.html"></a>
-        <a id="nextFileLink" class="nav" href="z_c68eb0c7512457e4___init___py.html"></a>
+        <a id="prevFileLink" class="nav" href="z_a44f0ac069e85531_test_ingredient_py.html"></a>
+        <a id="nextFileLink" class="nav" href="z_a44f0ac069e85531___init___py.html"></a>
         <button type="button" class="button_prev_file" data-shortcut="["></button>
         <button type="button" class="button_next_file" data-shortcut="]"></button>
         <button type="button" class="button_show_hide_help" data-shortcut="?"></button>
