+++ htmlcov/status.json
@@ -1,1 +1,1 @@
-{"note":"This file is an internal implementation detail to speed up HTML report generation. Its format can change at any time. You might be looking for the JSON report: https://coverage.rtfd.io/cmd.html#cmd-json","format":5,"version":"7.6.9","globals":"5a732d02a626b79b64064e36887abd95","files":{"z_c68eb0c7512457e4___init___py":{"hash":"175e3bf678676fc912361fb3b5bc0240","index":{"url":"z_c68eb0c7512457e4___init___py.html","file":"praktikum\\__init__.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":0,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_c68eb0c7512457e4_bun_py":{"hash":"2042f255157a1a5d6a96e602204555ee","index":{"url":"z_c68eb0c7512457e4_bun_py.html","file":"praktikum\\bun.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":8,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_c68eb0c7512457e4_burger_py":{"hash":"5fd589d8401f9a0f066724a675105407","index":{"url":"z_c68eb0c7512457e4_burger_py.html","file":"praktikum\\burger.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":27,"n_excluded":0,"n_missing":0,"n_branches":4,"n_partial_branches":0,"n_missing_branches":0}}},"z_c68eb0c7512457e4_database_py":{"hash":"cba03020953215bccfb0a1deb2c407e1","index":{"url":"z_c68eb0c7512457e4_database_py.html","file":"praktikum\\database.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":21,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_c68eb0c7512457e4_ingredient_py":{"hash":"c1fa964ca484ba21b430aa07ca01755a","index":{"url":"z_c68eb0c7512457e4_ingredient_py.html","file":"praktikum\\ingredient.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":11,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_c68eb0c7512457e4_ingredient_types_py":{"hash":"5cd3afdf45a72b11c8b4cf8e0506fe73","index":{"url":"z_c68eb0c7512457e4_ingredient_types_py.html","file":"praktikum\\ingredient_types.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":2,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}}}}
+{"note":"This file is an internal implementation detail to speed up HTML report generation. Its format can change at any time. You might be looking for the JSON report: https://coverage.rtfd.io/cmd.html#cmd-json","format":5,"version":"7.6.9","globals":"a3e2e2b363fdc0f56bb263442b79dd84","files":{"z_a44f0ac069e85531___init___py":{"hash":"175e3bf678676fc912361fb3b5bc0240","index":{"url":"z_a44f0ac069e85531___init___py.html","file":"tests\\__init__.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":0,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_a44f0ac069e85531_test_bun_py":{"hash":"7f9de832f2e16a8de2ae6156999b1e24","index":{"url":"z_a44f0ac069e85531_test_bun_py.html","file":"tests\\test_bun.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":19,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_a44f0ac069e85531_test_burger_py":{"hash":"c3582df0013cdfcfc19a5debcf9fad25","index":{"url":"z_a44f0ac069e85531_test_burger_py.html","file":"tests\\test_burger.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":33,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_a44f0ac069e85531_test_database_py":{"hash":"e0bf216133c5dada6ffdf888f5ee288f","index":{"url":"z_a44f0ac069e85531_test_database_py.html","file":"tests\\test_database.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":26,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_a44f0ac069e85531_test_ingredient_py":{"hash":"13e831d3150f211a8bc421cd1c232237","index":{"url":"z_a44f0ac069e85531_test_ingredient_py.html","file":"tests\\test_ingredient.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":25,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}}}}
