+++ tests/test_burger.py
@@ -2,6 +2,7 @@
 import pytest
 
 from data import RECEIPT_TEMPLATE
+from helpers import expected_burger_price, set_burger_attrs
 from praktikum.burger import Burger
 
 
@@ -11,19 +12,6 @@ class TestBurger:
     def create_burger_object(self):
         '''Создание нового объекта класса «Burger» для каждого теста.'''
         self.burger = Burger()
-
-    def set_burger_attrs(self, *ingredients, bun=None):
-        '''Передача бургеру булочки и ингредиентов.'''
-        self.burger.bun = bun
-        self.burger.ingredients = list(ingredients)
-
-    @staticmethod
-    def expected_burger_price(*ingredients, bun):
-        '''Ожидаемая цена бургера.'''
-        price = bun.get_price() * 2
-        for ingredient in ingredients:
-            price += ingredient.get_price()
-        return price
 
     def test_init_attr_bun_is_none(self):
         '''Атрибут «bun» по умолчанию - None.'''
@@ -45,33 +33,39 @@ class TestBurger:
 
     def test_remove_ingredient_deletes_ingredient(self, mock_ingredient):
         '''Удаление ингредиента бургера.'''
-        self.set_burger_attrs(mock_ingredient)
+        self.burger = set_burger_attrs(self.burger, mock_ingredient)
         self.burger.remove_ingredient(index=0)
         assert self.burger.ingredients == []
 
     def test_move_ingredient_relocates_ingredient(self, mock_ingredient):
         '''Перемещение ингредиента бургера.'''
-        self.set_burger_attrs('ingredient', mock_ingredient)
+        self.burger = set_burger_attrs(
+            self.burger, 'ingredient', mock_ingredient
+        )
         self.burger.move_ingredient(index=1, new_index=0)
         assert self.burger.ingredients == [mock_ingredient, 'ingredient']
 
     def test_get_price_returns_burger_price(self, mock_bun, mock_ingredient):
         '''Получение цены бургера.'''
-        self.set_burger_attrs(mock_ingredient, bun=mock_bun)
+        self.burger = set_burger_attrs(
+            self.burger, mock_ingredient, bun=mock_bun
+        )
         assert (
             self.burger.get_price() ==
-            self.expected_burger_price(mock_ingredient, bun=mock_bun)
+            expected_burger_price(mock_ingredient, bun=mock_bun)
         )
 
     def test_get_receipt_returns_burger_receipt(
         self, mock_bun, mock_ingredient
     ):
         '''Получение рецепта бургера.'''
-        self.set_burger_attrs(mock_ingredient, bun=mock_bun)
+        self.burger = set_burger_attrs(
+            self.burger, mock_ingredient, bun=mock_bun
+        )
         data = {
             'bun_name': mock_bun.get_name(),
             'ingredient_type': mock_ingredient.get_type().lower(),
             'ingredient_name': mock_ingredient.get_name(),
-            'price': self.expected_burger_price(mock_ingredient, bun=mock_bun)
+            'price': expected_burger_price(mock_ingredient, bun=mock_bun)
         }
         assert self.burger.get_receipt() == RECEIPT_TEMPLATE.format(**data)
